Refactored codebase for modularity.
