+++ .gitignore
@@ -2,4 +2,3 @@ dist
 node_modules
 .env
 credentials.json
-test_output.json
