+++ package.json
@@ -18,18 +18,22 @@
   "dependencies": {
     "@types/node-cron": "^3.0.11",
     "discord.js": "^14.16.2",
-    "dotenv": "^16.4.5",
+    "dotenv": "^16.4.7",
     "googleapis": "^144.0.0",
     "node-cron": "^3.0.3",
     "node-fetch": "^3.3.2",
     "readline": "^1.3.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.23.0",
     "@types/node": "^22.5.5",
     "@types/tape": "^5.6.4",
     "csv-parser": "^3.0.0",
+    "eslint": "^9.23.0",
+    "globals": "^16.0.0",
     "nodemon": "^3.1.4",
     "tape": "^5.9.0",
-    "typescript": "^5.6.2"
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.28.0"
   }
 }
