+++ test_output.txt
@@ -1,0 +1,28 @@
+
+> enki@1.0.0 test
+> tape dist/tests/**/kings.test.js
+
+TAP version 13
+# Kings (Behe/Faf/Ada) Channel To DKP - 5 windows, pop msg
+ok 1 should be deeply equivalent
+# Nov05 Ada4
+ok 2 should be deeply equivalent
+# Nov06 Ada5
+ok 3 should be deeply equivalent
+# Nov12 Beh2 - Alt name (darkfarkee)
+ok 4 should be deeply equivalent
+# Nov13 Ada7 - Alt name (darkfarkee)
+ok 5 should be deeply equivalent
+# Nov14 Beg4 - x-out timing, alt name
+ok 6 should be deeply equivalent
+# Nov15 Ada2 - early termination no pop with x2 and x-out member
+ok 7 should be deeply equivalent
+# Nov21 Ada4 - x-out and x-in format
+ok 8 should be deeply equivalent
+
+1..8
+# tests 8
+# pass  8
+
+# ok
+
